add parent categories and overall improve the categories controller & views

+++ neMag/Controllers/CategoriesController.cs
@@ -23,14 +23,15 @@ namespace online_shop.Controllers
         }
         public ActionResult Show(int id)
         {
-            ViewBag.category = db.Categories.Find(id);
-            return View();
+            Category category = db.Categories.Find(id);
+            return View(category);
         }
 
         // GET
         public ActionResult Edit(int id)
         {
             Category category = db.Categories.Find(id);
+            ViewBag.categories = GetAllCategories(category);
             return View(category);
         }
         [HttpPut]
@@ -43,7 +44,13 @@ namespace online_shop.Controllers
                 if(TryUpdateModel(cat))
                 {
                     cat.Title = requestCat.Title;
+                    cat.ParentId = requestCat.ParentId;
                     cat.Description = requestCat.Description;
+                    cat.ParentId = requestCat.ParentId;
+                    if (requestCat.ParentId > 0)
+                    {
+                        cat.Parent = db.Categories.Find(requestCat.ParentId);
+                    }
                     db.SaveChanges();
                     TempData["message"] = "Categorie schimbată cu succes.";
                     return RedirectToAction("Index");
@@ -58,7 +65,7 @@ namespace online_shop.Controllers
         public ActionResult New()
         {
             Category cat = new Category();
-            ViewBag.modificare = "DA";
+            ViewBag.categories = GetAllCategories(null);
             return View(cat);
         }
         [HttpPost]
@@ -66,13 +73,16 @@ namespace online_shop.Controllers
         {
             try
             {
+                if (newCat.ParentId > 0)
+                {
+                    newCat.Parent = db.Categories.Find(newCat.ParentId);
+                }
                 db.Categories.Add(newCat);
                 db.SaveChanges();
                 TempData["message"] = "Categorie nouă adaugată cu succes.";
                 return RedirectToAction("Index");
             }catch(Exception e)
             {
-                ViewBag.modificare = "NU";
                 return View(newCat);    
             }
         }
@@ -81,10 +91,50 @@ namespace online_shop.Controllers
         public ActionResult Delete(int id)
         {
             Category cat = db.Categories.Find(id);
+            /*
+             * Categories can only be deleted if there are no items listed under them and if they
+             * don't have any subcategories. Possible improvement: if all subcategories are empty,
+             * delete them all at once when their parent is removed.
+             */
+            bool hasSubcateg = db.Categories.Where(c => c.ParentId == cat.CategoryId).Any();
+            if(hasSubcateg)
+            {
+                TempData["message"] = "Nu se poate șterge o categorie care are alte subcategorii în ierarhie.";
+                return RedirectToAction("Index");
+            }
+            bool hasProducts = db.Products.Where(p => p.CategoryId == cat.CategoryId).Any();
+            if(hasProducts)
+            {
+                TempData["message"] = "Nu se poate șterge o categorie care are produse în baza de date.";
+                return RedirectToAction("Index");
+            }
             db.Categories.Remove(cat);
             db.SaveChanges();
             TempData["message"] = "Categorie ștearsa cu succes.";
             return RedirectToAction("Index");
+        }
+
+        [NonAction]
+        private IEnumerable<SelectListItem> GetAllCategories(Category me)
+        {
+            var selectList = new List<SelectListItem>();
+            selectList.Add(new SelectListItem
+            {
+                Value = "-1",
+                Text = "nicio categorie părinte"
+            });
+            foreach (Category cat in (db.Categories).ToList())
+            {
+                if (cat != me)
+                {
+                    selectList.Add(new SelectListItem
+                    {
+                        Value = cat.CategoryId.ToString(),
+                        Text = cat.Title.ToString() // ToString might be redundant
+                    });
+                }
+            }
+            return selectList;
         }
     }
 }
